requirements.txt

dlib @ file:///E:/tensorflow_package/dlib-19.7.0-cp36-cp36m-win_amd64.whl
numpy==1.19.3
opencv-python==3.4.2.17

django~=3.1.3
djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0
pillow~=8.0.1